// react-frontend/src/api/models/admin-order-period-statistics.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AdminOrderStatisticLine } from './admin-order-statistic-line';

/**
 * 
 * @export
 * @interface AdminOrderPeriodStatistics
 */
export interface AdminOrderPeriodStatistics {
    /**
     * 
     * @type {string}
     * @memberof AdminOrderPeriodStatistics
     */
    'periodStart'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminOrderPeriodStatistics
     */
    'periodEnd'?: string;
    /**
     * 
     * @type {Array<AdminOrderStatisticLine>}
     * @memberof AdminOrderPeriodStatistics
     */
    'statistics'?: Array<AdminOrderStatisticLine>;
}