// react-frontend/src/api/models/bulk-order-status-update-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface BulkOrderStatusUpdateRequest
 */
export interface BulkOrderStatusUpdateRequest {
    /**
     * 
     * @type {string}
     * @memberof BulkOrderStatusUpdateRequest
     */
    'deliveryDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof BulkOrderStatusUpdateRequest
     */
    'status'?: StatusEnum;
    /**
     * 
     * @type {string}
     * @memberof BulkOrderStatusUpdateRequest
     */
    'previousStatus'?: PreviousStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof BulkOrderStatusUpdateRequest
     */
    'itemName'?: string;
}

export const StatusEnum = {
    PAID: 'PAID',
    IN_PROGRESS: 'IN_PROGRESS',
    IN_DELIVERY: 'IN_DELIVERY',
    DELIVERED: 'DELIVERED',
    CANCELLED: 'CANCELLED',
    REFUNDED: 'REFUNDED'
} as const;

export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
export const PreviousStatusEnum = {
    PAID: 'PAID',
    IN_PROGRESS: 'IN_PROGRESS',
    IN_DELIVERY: 'IN_DELIVERY',
    DELIVERED: 'DELIVERED',
    CANCELLED: 'CANCELLED',
    REFUNDED: 'REFUNDED'
} as const;

export type PreviousStatusEnum = typeof PreviousStatusEnum[keyof typeof PreviousStatusEnum];